Added successfully the getImage method
Finally succeeded in getting Yesod to get the image from the mongodb
database and return it to the browser.

I even tested uploading the file using the MDE editor and see the resulting
blog post. And tadaa!!! the lolcat appeared.

major milestone. Now we can add tests to nail the functionality down.